==> 1.stock_return_rate.ipynb <==
###1. calculate stock return rate & remove outlier

# timeSeries14~18까지 데이터 로드

df14=pd.read_csv("timeSeries14_name.csv")
df15=pd.read_csv("timeSeries15_name.csv")
df16=pd.read_csv("timeSeries16_name.csv")
df17=pd.read_csv("timeSeries17_name.csv")
df18=pd.read_csv("timeSeries18_name.csv")

# transfose data formation
df14['month']=df14['date'].astype("str").astype('datetime64').dt.month
df14['year']=df14['date'].astype("str").astype('datetime64').dt.year


df15['month']=df15['date'].astype("str").astype('datetime64').dt.month
df15['year']=df15['date'].astype("str").astype('datetime64').dt.year

df16['month']=df16['date'].astype("str").astype('datetime64').dt.month
df16['year']=df16['date'].astype("str").astype('datetime64').dt.year


df17['month']=df17['date'].astype("str").astype('datetime64').dt.month
df17['year']=df17['date'].astype("str").astype('datetime64').dt.year


df18['month']=df18['date'].astype("str").astype('datetime64').dt.month
df18['year']=df18['date'].astype("str").astype('datetime64').dt.year

# col_name = stock name
# row = timeseries


col_name=df14.columns 

df15=df15[col_name]
df16=df16[col_name]
df17=df17[col_name]
df18=df18[col_name]

# conncat
df=pd.concat([df14,df15,df16,df17,df18])

# extract stock name 
stock_name=df.columns[1:-2]

# calcuate daily rate
rate=df[stock_name].pct_change()

# date return rate > 5 :  outlier (stock split issue)
# to do : Revised share price
outlier=[]
for i in rate.columns: # i = stock name
    for j in range(rate.shape[0]):
        if np.abs(rate[i].iloc[j])>5: # date return rate >5
            outlier.append(i)

# remove outlier
tmp=list(df.columns[1:])
for i in outlier:
    tmp.remove(i)
print(len(tmp)) # 1564 개 43개의 종목 제거
print(tmp)


# example : '코웨이' monthly average end-price
result=df[['year','month','코웨이']].groupby(['year','month']).mean()


# 모든 stock 한달 평균 종가 구하기
for name in stock_name[:-1]:
    print(name)
    tmp=df[['year','month',name]].groupby(['year','month']).mean()
    result=result.merge(tmp, how='inner', left_on=['year','month'], right_on=['year','month'])

    
result=result.drop(columns=['코웨이_y'])
result=result.rename(columns={'코웨이_x':'코웨이'})

# save
result.to_csv("./result/avg_price_month_delOutlier.csv")
