==> 2.sector_return_rate.ipynb <==
### 2. calculate sector return rate

# returnsector.csv : KRX sector data + return 1418 
df=pd.read_csv("returnSector.csv") 
df.to_csv("returnSector.csv",encoding='utf-8-sig')

# remove stock name which is not in KRX dataset
delName=[]
for name in df['기업명']:
    if name in outlier:
        print(name)
        delName.append(name) # 34개

tmp=list(df['기업명'])

for i in delName:
    if i in tmp:
        tmp.remove(i)
        
        
print("numbers of removed stock name",len(tmp)-len((df['기업명'])))

# extract index to be removed
indexdel=[]
for i in delName:
   
    #print(df[df['기업명']==i].index[0])
    indexdel.append(df[df['기업명']==i].index[0])
    
print(df.drop(indexdel))

# remove
df=df.drop(indexdel)

# check number of sector
len(df['업종'].unique()) # 140 sector

# 2014년 1월 데이터 nan 값 제거(1로 처리하였음)
# 필요없는 columns 제거
sector=sector.drop(columns=['종목코드','0','업종코드','Unnamed: 0','Unnamed: 0.1','Unnamed: 0.1.1','Unnamed: 0.1.1.1','Unnamed: 0.1.1.1.1','Unnamed: 0.1.1.1.1.1'])


# sector return rate
sector.to_csv("Sector_return_outlier.csv",encoding='utf-8-sig')

sector=sector.drop(columns=['Unnamed: 0.1.1'])

# calculate correlation among sectors
sector_corr=sector.T.corr()

# heatmap 
import seaborn as sns
import matplotlib.pyplot as plt
plt.figure(figsize=(10, 10))
sns.heatmap(sector_corr)

# sector : corr >0.7 
sns.heatmap(sector_corr[sector_corr>0.7].loc[sector_corr[sector_corr>0.7].columns])
