==> 3. calc_corr.ipynb <==
### 3. calculate sector correlation

from scipy.stats import pearsonr
dataset=sector.T
index=dataset.columns # 업종이름
index

table=[]

# 0.7이상의 상관관계를 가지는 업종짝 구하기
for i in range(len(index)):
    for j in range(len(index)):
        if (pearsonr(dataset[index[i]],dataset[index[j]])[0]>0.7)&(pearsonr(dataset[index[i]],dataset[index[j]])[1]<0.05)&(pearsonr(dataset[index[i]],dataset[index[j]])[0]!=1):
            # 상관계수가 0.75이고 p_value가 0.05보다 작은 것 고르기
            table.append((index[i],index[j],pearsonr(dataset[index[i]],dataset[index[j]])[0],pearsonr(dataset[index[i]],dataset[index[j]])[1]))
            print(index[i],index[j],pearsonr(dataset[index[i]],dataset[index[j]])[0])

table=pd.DataFrame(table)
table.columns=['sector1','sector2','corr','p_value']

# save
table.to_csv("corrTable7_outlier.csv",encoding='utf-8-sig') 

# stock 
stock=pd.read_csv("returnSector.csv")

# index를 기업명으로 바꾸기
stock.index=stock['기업명']

# 업종별로 추출하기
stock=stock.drop(columns=['Unnamed: 0','0','업종코드','기업명','종목코드'])

stock=stock.drop(columns=['name'])

# 0.7이상의 상관관계를 가지는 모든 업종의 각 종목 상관관계 구하기
index=table['sector1'].unique() # index는 sector부분
table=[]

for i in range(len(index)):
    stock_name=stock[stock['업종']==index[i]].index
    for j in range(len(stock_name)):
        if (pearsonr(dataset[index[i]],stock.loc[stock_name[j]].T.iloc[:-1])[1]<0.05)&(pearsonr(dataset[index[i]],stock.loc[stock_name[j]].T.iloc[:-1])[0]!=1):
            # 상관계수가 0.75이고 p_value가 0.05보다 작은 것 고르기
            table.append((index[i],stock_name[j],pearsonr(dataset[index[i]],stock.loc[stock_name[j]].T.iloc[:-1])[0],pearsonr(dataset[index[i]],stock.loc[stock_name[j]].T.iloc[:-1])[1]))
            
            print(index[i],stock_name[j],pearsonr(dataset[index[i]],stock.loc[stock_name[j]].T.iloc[:-1])[0])

stock_sector=pd.DataFrame(table)
stock_sector.columns=['sector','stock','corr','p_value']
stock_sector.to_csv("해당업종전체종목상관관계_이상치.csv",encoding='utf-8-sig')
