==> 4.draw network.ipynb <==
### 4. draw sector network

# import module
import pandas as pd 
import numpy as np
import collections

import networkx as nx
import seaborn as sb
import matplotlib.pylab as plt
from pylab import rcParams

from scipy.stats import pearsonr

# read correlation file
sector=pd.read_csv("corrTable7_outlier.csv")
stock_sec=pd.read_csv("stock_sector_corr_outlier.csv")

sector=sector.drop(columns=['Unnamed: 0'])
stock_sec=stock_sec.drop(columns=['Unnamed: 0'])

sector.columns=['sector1','sector2','corr','p-value']

sec_net=nx.Graph()

from matplotlib import font_manager, rc
font_name = font_manager.FontProperties(fname="c:/Windows/Fonts/malgun.ttf").get_name()
rc('font', family=font_name)

# 업종간 네트워크 
# lable 새로 쓰기
plt.rc('font', family='AppleGothic')
for i in sector['sector1']:
    sec_net.add_node(i)
    
nx.draw(sec_net,with_labels = True, font_family='AppleGothic')
plt.show()


len(sec_net.node) # 36 중복없음
print(sec_net.degree) # 몇 개 연결되어있는지


for (i,j,corr) in np.array(sector[['sector1','sector2','corr']]):
    sec_net.add_edge(i,j,weight=1/corr*0.1,label=round(corr,4))# 길이 어떻게 늘리는지
        # weight가 적을수록 멀어짐
    
plt.figure(figsize=(15,15))

edge_labels =dict([((u, v), d['label']) 
                   for u, v, d in sec_net.edges(data=True)])
pos = nx.spring_layout(sec_net) 
#nx.draw_networkx_edges(sec_net,pos,edgelist=edge_labels)
#nx.draw_networkx_edge_labels(sec_net, pos, edge_labels= edge_labels)
nx.draw(sec_net,pos,with_labels = True, node_size=200,width=6,alpha=0.5,font_family=font_name)
nx.draw_networkx_edge_labels(sec_net,pos, edge_labels= edge_labels)

plt.show()

### 5. draw sector_stock network

# data loda

sector=pd.read_csv("corrTable7_outlier.csv")
stock_sec=pd.read_csv("stock_sector_corr_outlier.csv")
sector=sector.drop(columns=['Unnamed: 0'])
stock_sec=stock_sec.drop(columns=['Unnamed: 0'])
sector.columns=['sector1','sector2','corr','p-value']

sector['corr']=sector['corr']

# graph 선언
sec_net=nx.Graph()


# 상관관계를 반영하여 edge 생성
sec_edge=np.array(sector[['sector1','sector2',"corr"]])
sec_net.add_weighted_edges_from(sec_edge)

nx.draw(sec_net, node_size=10)

# dictionary sector :1 ,stock :0

b={}
for i in sector['sector1'].unique():
    b[i]=1
    
sec_val=b

stock_val={}
for i in stock_sec['stock'].unique():
    stock_val[i]=0

# blue: stock /red : sector
blue_edges=np.array(stock_sec[['sector','stock']])
red_edges=np.array(sector[['sector1','sector2']])

stock_sec[['sector','stock']]

tuples=[tuple(x) for x in stock_sec[['sector','stock']].values]

blue_edges=tuples

edge_colours = ['red' if not edge in blue_edges else 'blue'
                for edge in sec_net.edges()]

plt.figure(figsize=(15,15))
nx.draw_networkx_nodes(sec_net, pos, cmap=plt.get_cmap('jet'), 
                       node_color =values , node_size = 70)

nx.draw_networkx_labels(sec_net, pos, font_size=8, font_family=font_name)
nx.draw_networkx_edges(sec_net, pos, edgelist=blue_edges,edge_color='b')
nx.draw_networkx_edges(sec_net, pos, edgelist=red_edges)
